requirements.txt
absl-py==2.3.1
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
builtin-interfaces==1.2.2
cloudpickle @ file:///croot/cloudpickle_1751309439603/work
composition-interfaces==1.2.2
contourpy==1.3.0
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
etils==1.5.2
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
Farama-Notifications @ file:///croot/farama-notifications_1683058702908/work
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
geometry-msgs==4.9.0
glfw==2.9.0
grpcio==1.74.0
gymnasium==1.1.1
image-geometry==3.2.1
importlib_metadata @ file:///croot/importlib_metadata-suite_1732633488278/work
importlib_resources==6.5.2
interactive-markers==2.3.2
jax-jumpy @ file:///croot/jax-jumpy_1683230527192/work
Jinja2==3.1.6
kiwisolver==1.4.7
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.9.4
message-filters==4.3.8
mpmath==1.3.0
mujoco==3.3.4
nav-msgs==4.9.0
networkx==3.2.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1671484982936/work
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.3.0
protobuf==3.20.3
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sros2==0.10.7
stable_baselines3==2.7.0
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tf2-geometry-msgs==0.25.15
tf2-kdl==0.25.15
tf2-msgs==0.25.15
tf2-py==0.25.15
tf2-ros-py==0.25.15
tf2-sensor-msgs==0.25.15
tf2-tools==0.25.15
topic-monitor==0.20.5
torch==2.7.1
torchvision==0.22.1
trajectory-msgs==4.9.0
triton==3.3.1
turtlesim==1.4.2
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
unique-identifier-msgs==2.2.1
visualization-msgs==4.9.0
Werkzeug==3.1.3
zipp @ file:///croot/zipp_1732630741423/work
